CARDS-2654: Refactor Group and User related components to functional All files with a main component should follow CamelCaseNamingConvention
