+++ modules/principals/src/main/frontend/assets.config
@@ -2,5 +2,5 @@
 
 ['cards-principals.userIcon']: '@mui/icons-material/Person'
 ['cards-principals.groupIcon']: '@mui/icons-material/Group'
-['cards-principals.userscontainer']: './src/Userboard/Users/usersContainer.jsx'
-['cards-principals.groupscontainer']: './src/Userboard/Groups/groupsContainer.jsx'
+['cards-principals.UsersContainer']: './src/Userboard/Users/UsersContainer.jsx'
+['cards-principals.GroupsContainer']: './src/Userboard/Groups/GroupsContainer.jsx'
