+++ modules/principals/src/main/frontend/src/Userboard/Users/UsersManager.jsx
@@ -21,9 +21,9 @@ import { checkPropTypes } from "../../propTypes";
 import { withStyles } from 'tss-react/mui';
 import { Avatar, Box, Card, CardContent, IconButton, Tooltip } from "@mui/material";
 import userboardStyle from '../userboardStyle.jsx';
-import CreateUserDialog from "./createuserdialog.jsx";
-import DeletePrincipalDialog from "../deleteprincipaldialog.jsx";
-import ChangeUserPasswordDialog from "./changeuserpassworddialog.jsx";
+import CreateUserDialog from "./CreateUserDialog.jsx";
+import DeletePrincipalDialog from "../DeletePrincipalDialog.jsx";
+import ChangeUserPasswordDialog from "./ChangeUserPasswordDialog.jsx";
 import NewItemButton from "../../components/NewItemButton.jsx";
 import AdminScreen from "../../adminDashboard/AdminScreen.jsx";
 import MaterialReactTable from 'material-react-table';
