+++ modules/principals/src/main/resources/SLING-INF/content/Extensions/Views/Grouplist.json
@@ -3,5 +3,5 @@
   "cards:extensionPointId": "cards/coreUI/view",
   "cards:extensionName": "Groups",
   "cards:targetURL": "/content.html/admin/grouplist",
-  "cards:extensionRenderURL": "asset:cards-principals.groupscontainer.js"
+  "cards:extensionRenderURL": "asset:cards-principals.GroupsContainer.js"
 }
